requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.29
pymysql==1.1.0
python-dotenv==1.1.0
alembic==1.13.1
pydantic==2.11.0
pytest==8.1.1
attrs==23.2.0
azure-storage-blob==12.19.0
Jinja2==3.1.3
cryptography==44.0.2
gunicorn==21.2.0
azure-cosmos==4.6.0
botbuilder-core
botbuilder-dialogs
aiohttp
requests

